Rename dir from util to hooks
